Fix the "release" stage in .travis.yml
Travis CI acts weird.
For each tagged commit in master two separate builds are triggered:
one "for tag" and the other "for branch".
However, the "for tag" build has the tag value as its branch name,
while the "for branch" build has no tags at all.
The only safe way to auto-release packages is to skip "for tag" builds,
and release on every push to master (presuming it's tagged).