Stop using pickling for serialization
Pickling is a simple but very limiting approach to serialization in Python
and needs to be replaced with some stable serialization method.

However, popular serialization libraries (like `marshmallow`) do not support
all required Python versions (like Python 3.4 or PyPy).
So, we had to implement some simple stuff on our own.